requirements.txt
Flask==2.2.2
gunicorn==20.0.4
numpy==1.24.0
opencv_python==4.6.0.66
pandas==1.5.2
pdf2image==1.16.0
Pillow==9.4.0
pytesseract==0.3.10
